Update PythiaGun.cc
The straightforward updates have been made to allow 100 runs of every pthat bin (33 in this case) to be conducted on the same concurrently generated hydro profile through MUSIC. In concurrent runs, hydro profiles are generated on the fly. The reuse hydro tag allows running multiple events on the same pthat bin, but running different pthat bins on the same hydro profile wasn't possible before. This update to the Pythia gun addresses that issue. The final output file will contain sets of 100 events from the corresponding pthat bin.

Another solution would have been to set the same seed and expect the same hydro profile to be generated if the seed is consistent, but I wasn't sure if that approach would be reliable. Some tests would be needed to confirm its effectiveness.