Update Matter.cc
Few lines are added to matter code so that the virtuality scale where splitting happens in tracked with time. The purpose was to check how elastic collision affects the virtuality and if it is ok to decide the virtuality scale by doing elastic collision. There is a major conflict in this because the, split time is decided by virtuality, so in some way there is pre-determination going on, when you say there will be a split and we are changing that pre-detrmined fact by defining a new virtuality scale.